Project Start
Spa system project start. Created the register and login system. Enquiries are working with a mailto and the front end for creating a booking is working. Applied bootstrap theme. Authorisation has been added for the create booking page so a user must have an account to get to the create booking page